requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-chess>=1.9
